// server/services/interfaces/corePerson/corePersonPhysicalAttributes.ts
export interface CorePersonPhysicalAttributes {
  height?: number
  weight?: number
  hairCode?: string
  hairDescription?: string
  facialHairCode?: string
  facialHairDescription?: string
  faceCode?: string
  faceDescription?: string
  buildCode?: string
  buildDescription?: string
  leftEyeColourCode?: string
  leftEyeColourDescription?: string
  rightEyeColourCode?: string
  rightEyeColourDescription?: string
  shoeSize?: string
}